+++ src/main/java/jsonvalues/OpUnionJsons.java
@@ -2,9 +2,7 @@ package jsonvalues;
 
 
 class OpUnionJsons {
-    // squid:S1452: Json<?> has only two possible types: JsObj or JsArr,
-    // squid:S00117: ARRAY_AS  should be a valid name
-    @SuppressWarnings({"squid:S1452", "squid:S00117"})
+
     static Json<?> unionAll(Json<?> a,
                             Json<?> b,
                             JsArray.TYPE ARRAY_AS